Fix timing issue
This line of code as written creates a TIH timing of 625ns, which is short of the 800ns recommended. Some neopixels will not work with this timing. An example of the problem and the solution (as identified and solved by Twitter user @Bornach) can be found in this Twitter thread (https://twitter.com/Bornach1/status/1301087477034344449?s=20)